This commit does not consist every change. But the highlights are: (#2)
- I added a code that detects note from audio (but it can detect only one note at a time... I'm working on a way to improve this)
- I cleaned and made sure audio_process.py will match to this format
- I added branches

It's kind of difficult to monitor all of the changes made... I haven't committed in a long time, because I was working and test-benching a lot of things. And since I'm kinda lazy and need to study (and so, I don't work continously on this project), I didn't make a comparison between the current push and the last one.

I hope that I'll work more clean from now on. I'll make sure to test stuff only in the TB branch, and work on stuff only in the correct branch.